refactor(*): make mapping classes generic
We need to be able to re-map the TypeScript declaration files as well as the generated ECMAscript
output files. This will require almost exactly the same process as the ECMAscript re-mapping. The
classes that handle the re-mapping have been made generic so that we can re-use the same framework
for the declaration re-mapping. See #2

BREAKING CHANGE: The re-mapping classes have been made generic and there are ES derived variants of
each. This is required so that we can nicely re-map TypeScript declaration files using the same
re-mapping framework. The demo library code in the `README` will still work, however, the modules
have all moved around hence the breaking change :boom: If you do not do any module loading of the
library, then you will require no code changes other than upgrading your dependency