ci(circle): upload greenkeeper lockfile in first job
The upload will fail with a `Only running on first push of new branch`
error if `CIRCLE_PREVIOUS_BUILD_NUM` is defined. That variable is
defined for every subsequent job. We need to run the lockfile update
in the first job. Luckily we have a job that runs at the start and
everything depends on it.

See https://github.com/greenkeeperio/greenkeeper-lockfile/issues/50